update rovio.info, merged test_rovio/rovio_node into one
