+++ cfg/rovio.info
@@ -63,11 +63,11 @@ Init
         gyb_0 3e-4;			X-Covariance of initial gyroscope bias [rad^2/s^2]
         gyb_1 3e-4;			Y-Covariance of initial gyroscope bias [rad^2/s^2]
         gyb_2 3e-4;			Z-Covariance of initial gyroscope bias [rad^2/s^2]
-        vep 1;				Covariance of initial linear camera-IMU extrinsics, same for all entries [m^2]
+        vep 0.0001;			Covariance of initial linear camera-IMU extrinsics, same for all entries [m^2]
         att_0 0.1;			X-Covariance of initial attitude [rad^2]
         att_1 0.1;			Y-Covariance of initial attitude [rad^2]
         att_2 0.1;			Z-Covariance of initial attitude [rad^2]
-        vea 0.1;			Covariance of initial rotational camera-IMU extrinsics, same for all entries [rad^2]
+        vea 0.0001;			Covariance of initial rotational camera-IMU extrinsics, same for all entries [rad^2]
     }
 }
 ImgUpdate
@@ -79,11 +79,11 @@ ImgUpdate
     startLevel 3;												Highest patch level which is being employed (must be smaller than the hardcoded template parameter)
     endLevel 1;													Lowest patch level which is being employed
     nDetectionBuckets 100;										Number of discretization buckets used during the candidates selection
-    MahalanobisTh 5.8858356;									Mahalanobis treshold for the update
+    MahalanobisTh 0.1;											Mahalanobis treshold for the update, 5.8858356
     UpdateNoise
     {
         nor 1e-5;												Covariance used for the reprojection error, 1/focal_length is roughly 1 pixel std [rad] (~1/f ~ 1/400^2 = 1/160000)
-        int 1e3;												Covariance used for the photometric error [intensity^2]
+        int 100;												Covariance used for the photometric error [intensity^2]
     }
     initCovFeature_0 0.5;										Covariance for the initial distance (in the selected parametrization)
     initCovFeature_1 1e-5;										Covariance for the initial bearing vector, x-component [rad]
@@ -98,19 +98,19 @@ ImgUpdate
     statMinGlobalQualityRange 100;								Minimum number of frames for obtaining maximal global quality
     trackingUpperBound 0.9;										Threshold for local quality for min overall global quality
     trackingLowerBound 0.1;										Threshold for local quality for max overall global quality
-    minTrackedAndFreeFeatures 0.5;								Minimum of amount of feature which are either tracked or free
+    minTrackedAndFreeFeatures 0.75;								Minimum of amount of feature which are either tracked or free
     removalFactor 1.1;											Factor for enforcing feature removal if not enough free
     minRelativeSTScore 0.75;									Minimum relative ST-score for extracting new feature patch
     minAbsoluteSTScore 5.0;										Minimum absolute ST-score for extracting new feature patch
     minTimeBetweenPatchUpdate 1.0;								Minimum time between new multilevel patch extrection [s]
-    fastDetectionThreshold 10;									Fast corner detector treshold while adding new feature
+    fastDetectionThreshold 5;									Fast corner detector treshold while adding new feature
     alignConvergencePixelRange 1;								Assumed convergence range for image alignment (gets scaled with the level) [pixels]
     alignCoverageRatio 2;										How many sigma of the uncertainty should be covered in the adaptive alignement
-    alignMaxUniSample 3;										Maximal number of alignment seeds on one side -> total number of sample = 2n+1. Carefull can get very costly if diverging!
+    alignMaxUniSample 1;										Maximal number of alignment seeds on one side -> total number of sample = 2n+1. Carefull can get very costly if diverging!
     matchingPixelThreshold 0.0;									Below this treshold no pre-alignment is carried out [pixels]
     minNoAlignment 5;											Minimal number of alignment every feature must make through.
-    specialLinearizationThreshold 0.0025;						Below this treshold no special linearization point gets computed (1 pixel ~ 0.0025) [rad]
-    patchRejectionTh 10;										If the average itensity error is larger than this the feauture is rejected [intensity]
+    specialLinearizationThreshold 0.0;							Below this treshold no special linearization point gets computed (1 pixel ~ 0.0025) [rad]
+    patchRejectionTh -1;										If the average itensity error is larger than this the feauture is rejected [intensity], if smaller 0 the no check is performed
     bearingVectorMahalTh 9.21;									Threshold for the aligned patch to be accepted as inlier
     removeNegativeFeatureAfterUpdate true;						Should feature with negative distance get removed
     maxUncertaintyToDepthRatioForDepthInitialization 0.3;		If set to 0.0 the depth is initialized with the standard value provided above, otherwise ROVIO attempts to figure out a median depth in each frame
@@ -118,7 +118,7 @@ ImgUpdate
     doStereoInitialization true;								Should a stereo match be used for feature initialization.
     MotionDetection
     {
-    	isEnabled 1;											Is the motion detection enabled
+    	isEnabled 0;											Is the motion detection enabled
 	    rateOfMovingFeaturesTh 0.5;								Amount of feature with motion for overall motion detection
 	    pixelCoordinateMotionTh 1.0;							Threshold for motion detection for patched [pixels]
 	    minFeatureCountForNoMotionDetection 5;					Min feature count in frame for motion detection
@@ -133,7 +133,7 @@ ImgUpdate
         }
         MahalanobisTh0 7.689997599999999;						Mahalanobid distance for zero velocity updates
         minNoMotionTime 1.0;									Min duration of no-motion
-        isEnabled 1;											Should zero velocity update be applied, only works if MotionDetection.isEnabled is true
+        isEnabled 0;											Should zero velocity update be applied, only works if MotionDetection.isEnabled is true
     }
 }
 Prediction
@@ -143,9 +143,9 @@ Prediction
         pos_0 1e-4;								X-covariance parameter of position prediction [m^2/s]
         pos_1 1e-4;								Y-covariance parameter of position prediction [m^2/s]
         pos_2 1e-4;								Z-covariance parameter of position prediction [m^2/s]
-        vel_0 4e-6;								X-covariance parameter of velocity prediction [m^2/s^3]
-        vel_1 4e-6;								Y-covariance parameter of velocity prediction [m^2/s^3]
-        vel_2 4e-6;								Z-covariance parameter of velocity prediction [m^2/s^3]
+        vel_0 4e-5; 4e-6						X-covariance parameter of velocity prediction [m^2/s^3]
+        vel_1 4e-5;								Y-covariance parameter of velocity prediction [m^2/s^3]
+        vel_2 4e-5;								Z-covariance parameter of velocity prediction [m^2/s^3]
         acb_0 1e-8;								X-covariance parameter of accelerometer bias prediction [m^2/s^5]
         acb_1 1e-8;								Y-covariance parameter of accelerometer bias prediction [m^2/s^5]
         acb_2 1e-8;								Z-covariance parameter of accelerometer bias prediction [m^2/s^5]
@@ -153,9 +153,9 @@ Prediction
         gyb_1 3.8e-7;							Y-covariance parameter of gyroscope bias prediction [rad^2/s^3]
         gyb_2 3.8e-7;							Z-covariance parameter of gyroscope bias prediction [rad^2/s^3]
         vep 1e-6;								Covariance parameter of linear extrinsics prediction [m^2/s]
-        att_0 7.6e-7;							X-Covariance parameter of attitude prediction [rad^2/s]
-        att_1 7.6e-7;							Y-Covariance parameter of attitude prediction [rad^2/s]
-        att_2 7.6e-7;							Z-Covariance parameter of attitude prediction [rad^2/s]
+        att_0 7.6e-6; 7.6e-7					X-Covariance parameter of attitude prediction [rad^2/s]
+        att_1 7.6e-6;							Y-Covariance parameter of attitude prediction [rad^2/s]
+        att_2 7.6e-6;							Z-Covariance parameter of attitude prediction [rad^2/s]
         vea 1e-8;								Covariance parameter of rotational extrinsics prediction [rad^2/s]
         dep 0.000001;							Covariance parameter of distance prediction [m^2/s]
         nor 0.0001;								Covariance parameter of bearing vector prediction [rad^2/s]
